first batch of PHP 8 fixes

+++ classes/manage.class.php
@@ -105,16 +105,16 @@ class MailsterManage {
 
 		global $wpdb;
 
-		$memory_limit       = @ini_get( 'memory_limit' );
-		$max_execution_time = @ini_get( 'max_execution_time' );
+		$memory_limit       = ini_get( 'memory_limit' );
+		$max_execution_time = ini_get( 'max_execution_time' );
 
-		@set_time_limit( 0 );
+		set_time_limit( 0 );
 
 		if ( (int) $max_execution_time < 300 ) {
-			@ini_set( 'max_execution_time', 300 );
+			ini_set( 'max_execution_time', 300 );
 		}
 		if ( (int) $memory_limit < 256 ) {
-			@ini_set( 'memory_limit', '256M' );
+			ini_set( 'memory_limit', '256M' );
 		}
 
 		if ( isset( $_FILES['async-upload'] ) ) {
@@ -137,9 +137,7 @@ class MailsterManage {
 
 			if ( ! current_user_can( 'mailster_import_subscribers' ) ) {
 
-				@header( 'Content-type: application/json' );
-				echo json_encode( $return );
-				exit;
+				wp_send_json( $return );
 			}
 
 			$raw_data = esc_textarea( stripslashes( $_POST['data'] ) );
@@ -150,9 +148,7 @@ class MailsterManage {
 
 			if ( ! current_user_can( 'mailster_import_wordpress_users' ) ) {
 
-				@header( 'Content-type: application/json' );
-				echo json_encode( $return );
-				exit;
+				wp_send_json( $return );
 			}
 
 			parse_str( $_POST['wordpressusers'], $data );
@@ -277,9 +273,7 @@ class MailsterManage {
 
 		if ( isset( $return ) ) {
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 	}
@@ -295,9 +289,7 @@ class MailsterManage {
 
 		if ( ! current_user_can( 'mailster_import_subscribers' ) ) {
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		$return['identifier'] = $identifier = $_POST['identifier'];
@@ -459,9 +451,7 @@ class MailsterManage {
 
 		$return['html'] = $html;
 
-		@header( 'Content-type: application/json' );
-		echo json_encode( $return );
-		exit;
+		wp_send_json( $return );
 
 	}
 
@@ -472,18 +462,18 @@ class MailsterManage {
 
 		define( 'MAILSTER_DO_BULKIMPORT', true );
 
-		$memory_limit       = @ini_get( 'memory_limit' );
-		$max_execution_time = @ini_get( 'max_execution_time' );
+		$memory_limit       = ini_get( 'memory_limit' );
+		$max_execution_time = ini_get( 'max_execution_time' );
 
-		@ini_set( 'display_errors', 0 );
+		ini_set( 'display_errors', 0 );
 
-		@set_time_limit( 0 );
+		set_time_limit( 0 );
 
 		if ( (int) $max_execution_time < 300 ) {
-			@ini_set( 'max_execution_time', 300 );
+			ini_set( 'max_execution_time', 300 );
 		}
 		if ( (int) $memory_limit < 256 ) {
-			@ini_set( 'memory_limit', '256M' );
+			ini_set( 'memory_limit', '256M' );
 		}
 
 		$return['success'] = false;
@@ -491,9 +481,7 @@ class MailsterManage {
 		$this->ajax_nonce( json_encode( $return ) );
 
 		if ( ! current_user_can( 'mailster_import_subscribers' ) ) {
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		$timeoffset = mailster( 'helper' )->gmt_offset( true );
@@ -557,7 +545,7 @@ class MailsterManage {
 						continue;
 					}
 
-					@set_time_limit( 10 );
+					set_time_limit( 10 );
 
 					$data       = explode( $bulkdata['separator'], $line );
 					$line_count = count( $data );
@@ -778,9 +766,7 @@ class MailsterManage {
 		}
 		$return['success'] = true;
 
-		@header( 'Content-type: application/json' );
-		echo json_encode( $return );
-		exit;
+		wp_send_json( $return );
 	}
 
 
@@ -794,9 +780,7 @@ class MailsterManage {
 		if ( ! current_user_can( 'mailster_export_subscribers' ) ) {
 			$return['msg'] = esc_html__( 'You are not allowed to export subscribers!', 'mailster' );
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		parse_str( $_POST['data'], $d );
@@ -864,9 +848,7 @@ class MailsterManage {
 			$return['msg'] = esc_html__( 'No Subscribers found!', 'mailster' );
 		}
 
-		@header( 'Content-type: application/json' );
-		echo json_encode( $return );
-		exit;
+		wp_send_json( $return );
 
 	}
 
@@ -882,9 +864,7 @@ class MailsterManage {
 		if ( ! current_user_can( 'mailster_export_subscribers' ) ) {
 			$return['msg'] = esc_html__( 'You are not allowed to export subscribers!', 'mailster' );
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		$filename = get_option( 'mailster_export_filename' );
@@ -892,9 +872,7 @@ class MailsterManage {
 		if ( ! file_exists( $filename ) || ! wp_is_writable( $filename ) ) {
 			$return['msg'] = esc_html__( 'Not able to write export file', 'mailster' );
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		parse_str( $_POST['data'], $d );
@@ -1174,11 +1152,13 @@ class MailsterManage {
 			} else {
 				$return['finished'] = true;
 
-				$finalname         = MAILSTER_UPLOAD_DIR . '/mailster_export_' . date( 'Y-m-d-H-i-s' ) . '.' . $outputformat;
-				$return['success'] = copy( $filename, $finalname );
-				$file_size         = @filesize( $filename );
-				update_option( 'mailster_export_filename', $finalname );
-				@unlink( $filename );
+				$finalname = MAILSTER_UPLOAD_DIR . '/mailster_export_' . date( 'Y-m-d-H-i-s' ) . '.' . $outputformat;
+				if ( file_exists( $filename ) ) {
+					$return['success'] = copy( $filename, $finalname );
+					$file_size         = filesize( $filename );
+					update_option( 'mailster_export_filename', $finalname );
+					unlink( $filename );
+				}
 				$return['filename'] = admin_url( 'admin-ajax.php?action=mailster_download_export_file&file=' . basename( $finalname ) . '&format=' . $outputformat . '&_wpnonce=' . wp_create_nonce( 'mailster_nonce' ) );
 			}
 
@@ -1191,9 +1171,7 @@ class MailsterManage {
 
 		}
 
-		@header( 'Content-type: application/json' );
-		echo json_encode( $return );
-		exit;
+		wp_send_json( $return );
 	}
 
 
@@ -1271,9 +1249,7 @@ class MailsterManage {
 		if ( ! current_user_can( 'mailster_bulk_delete_subscribers' ) ) {
 			$return['msg'] = 'no allowed';
 
-			@header( 'Content-type: application/json' );
-			echo json_encode( $return );
-			exit;
+			wp_send_json( $return );
 		}
 
 		parse_str( $_POST['data'], $d );
@@ -1332,9 +1308,7 @@ class MailsterManage {
 			$return['msg'] = esc_html__( 'No Subscribers removed', 'mailster' );
 		}
 
-		@header( 'Content-type: application/json' );
-		echo json_encode( $return );
-		exit;
+		wp_send_json( $return );
 
 	}
